// types/appointment.ts
// types/appointment.ts
export interface Appointment {
  id: number;
  date: string;
  age: string;
  time: string;
  patientName: string;
  issue: string;
  location: string;
  bloodGroup: string;
  height: string;
  weight: string;
  speciality: string;
  consultationType: string;
  appointmentDate: string;
  appointmentTime: string;
  allergies: string[];
}